Completed Long Video Upload in mobile section
