+++ app/upload/_components/shortUpload.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+
+
+const shortUpload = () => {
+  return (
+    <div>shortUpload</div>
+  )
+}
